feat: drop bellow 7.4, add support for symfony 6
- removed support to php versions before 7.4
- removed support to symfony before 4
- removed symfony/debug dependence (abandoned)
- add support to symfony 6 (symfony/lock has changed)
- AbstractFileStorage is needed to keep symfony 4.0 support
