+++ .gitignore
@@ -47,3 +47,5 @@
 # Embedded web-server pid file
 /.web-server-pid
 Tests/cache/
+
+composer.lock
