+++ Controller/PingAction.php
