+++ DependencyInjection/SimonettiPingExtension.php
@@ -1,10 +1,12 @@
 <?php
 namespace Simonetti\PingBundle\DependencyInjection;
 
+use Simonetti\PingBundle\Lock\FileStorage;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Extension\Extension;
 use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\Lock\Factory;
 
 /**
  * Class PingExtension
@@ -22,6 +24,14 @@ class SimonettiPingExtension extends Extension
 
         $definition = $container->getDefinition('simonetti.bundle.ping.lock');
         $definition->replaceArgument(0, $config['lock_name']);
+
+        if (class_exists(Factory::class)) {
+            $definition = $container->getDefinition('simonetti.bundle.ping.lock_factory');
+            $definition->setClass(Factory::class);
+
+            $definition = $container->getDefinition('simonetti.bundle.ping.lock_store');
+            $definition->setClass(FileStorage::class);
+        }
     }
 
 }
